changed to use Eventlet and doubles; updates #32'

+++ prototype2/gdgem/nmx/Clusterer.cpp
@@ -7,19 +7,19 @@
 //#undef TRC_LEVEL
 //#define TRC_LEVEL TRC_L_DEB
 
-Clusterer::Clusterer(uint64_t min_time_span) : min_time_span_(min_time_span) {}
+Clusterer::Clusterer(double min_time_span) : min_time_span_(min_time_span) {}
 
 void Clusterer::insert(const Eventlet &eventlet) {
   if ((eventlet.time < latest_time_) &&
       ((latest_time_ - eventlet.time) >
-       (std::numeric_limits<uint64_t>::max() / 2))) {
+       (std::numeric_limits<double>::max() / 2))) {
     XTRACE(PROCESS, ALW, "Clock overflow event %" PRIu64 " < %" PRIu64
                          " && %" PRIu64 " > %" PRIu64,
            eventlet.time, latest_time_, (eventlet.time - latest_time_),
-           (std::numeric_limits<uint64_t>::max() / 2));
+           (std::numeric_limits<double>::max() / 2));
     current_time_offset_ = latest_time_;
   }
-  backlog_.insert(std::pair<uint64_t, Eventlet>(
+  backlog_.insert(std::pair<double, Eventlet>(
       current_time_offset_ + eventlet.time, eventlet));
   latest_time_ = std::max(latest_time_, eventlet.time);
 }
